Moved new place page i18n test out of shared testing folder to avoid CDC test failures
