+++ server/webdriver/shared_tests/place_explorer_i18n_test.py
@@ -20,13 +20,6 @@ from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
 from server.webdriver import shared
-from server.webdriver.base_utils import find_elem
-
-# Regular expression for matching Japanese characters:
-# Hiragana: \u3040-\u309F
-# Katakana: \u30A0-\u30FF
-# Common Kanji: \u4E00-\u9FAF (includes many Chinese characters used in Japanese)
-JAPANESE_CHAR_PATTERN = re.compile(r'[\u3040-\u30FF\u4E00-\u9FAF]')
 
 
 class PlaceI18nExplorerTestMixin():
@@ -72,91 +65,6 @@ class PlaceI18nExplorerTestMixin():
     aichi_prefecture = self.driver.find_element(
         By.XPATH, '//*[@id="child-place"]/div/a[1]')
     self.assertEqual(aichi_prefecture.text, '三重県,')
-
-    # Test that timeline links are removed
-    self.assertListEqual(
-        self.driver.find_elements(By.CLASS_NAME, 'explore-more'), [])
-
-  def test_japan_in_japanese_new_place_page(self):
-    """Test translations from various sources are displayed correctly."""
-
-    start_url = self.url_ + '/place/country/JPN?hl=ja&force_dev_places=true'
-    self.driver.get(start_url)
-
-    place_name_present = EC.text_to_be_present_in_element(
-        (By.CSS_SELECTOR, '.place-info [data-testid="place-name"]'), '日本')
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(place_name_present)
-
-    # TODO: Update this test once the place type is translated
-    # Ensure that the place type in {parentPlace} is translated
-    place_type_present = EC.text_to_be_present_in_element(
-        (By.CSS_SELECTOR, '.place-info .subheader'), '国 in')
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(place_type_present)
-
-    # Ensure that the topics tab links are translated
-    economics_link = find_elem(self.driver,
-                               by=By.CSS_SELECTOR,
-                               value=".explore-topics-box .item-list-item a")
-    self.assertEqual(economics_link.text, "経済")
-
-    # Ensure that the economics section is translated
-    # The first block-title-text is the economics section
-    economics_section_present = EC.text_to_be_present_in_element(
-        (By.CLASS_NAME, 'block-title-text'), '経済')
-    WebDriverWait(self.driver,
-                  self.TIMEOUT_SEC).until(economics_section_present)
-
-    # Test that charts are present
-    charts = self.driver.find_elements(By.CSS_SELECTOR, '.chart-container')
-    self.assertGreater(len(charts), 0,
-                       "Expected at least one chart to be present")
-
-    # Wait for and scroll to ranking tile so it lazy loads
-    ranking_tile_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, "ranking-tile"))
-    ranking_tile = WebDriverWait(self.driver,
-                                 self.TIMEOUT_SEC).until(ranking_tile_present)
-    self.driver.execute_script("arguments[0].scrollIntoView();", ranking_tile)
-
-    # Ensure the ranking tile footer text is translated
-    self.assertEqual(
-        find_elem(self.driver,
-                  by=By.CSS_SELECTOR,
-                  value=".chart-container .chart-footnote").text,
-        "使用可能な最新のデータに基づくランキング。一部の地域は、対象の年の報告が不完全なため、欠落している可能性があります。")
-
-    # Wait for and scroll to the first bar chart tile so it lazy loads
-    bar_chart_tile_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, "bar-chart"))
-    bar_chart_tile = WebDriverWait(
-        self.driver, self.TIMEOUT_SEC).until(bar_chart_tile_present)
-    self.driver.execute_script("arguments[0].scrollIntoView();", bar_chart_tile)
-
-    # Ensure that the bar chart's axis labels (localized place names) have japanese text
-    self.assertTrue(
-        JAPANESE_CHAR_PATTERN.search(
-            find_elem(self.driver,
-                      by=By.CSS_SELECTOR,
-                      value="svg text.place-tick").text))
-
-    # TODO: Update this test once the see per capita link is translated
-    self.assertEqual(
-        find_elem(self.driver,
-                  by=By.CSS_SELECTOR,
-                  value=".block-controls [data-testid='see-per-capita']").text,
-        "See per capita")
-
-    # Test that related places callout is translated
-    related_places = find_elem(self.driver,
-                               by=By.CLASS_NAME,
-                               value="related-places-callout")
-    self.assertEqual(related_places.text, "日本 の地域")
-
-    # Test that child place link is translated
-    child_place_link = find_elem(self.driver,
-                                 by=By.CSS_SELECTOR,
-                                 value=".related-places .item-list-item a")
-    self.assertEqual(child_place_link.text, "北海道")
 
     # Test that timeline links are removed
     self.assertListEqual(
